feat: send early calibration performance stats
