+++ packages/client/src/rtc/BasePeerConnection.ts
@@ -8,19 +8,13 @@ import { createSafeAsyncSubscription } from '../store/rxUtils';
 import {
   ClientDetails,
   PeerType,
-  PerformanceStats,
   TrackType,
 } from '../gen/video/sfu/models/models';
 import { StreamSfuClient } from '../StreamSfuClient';
 import { AllSfuEvents, Dispatcher } from './Dispatcher';
 import { withoutConcurrency } from '../helpers/concurrency';
-import {
-  deltaCompression,
-  toObject,
-  Tracer,
-  traceRTCPeerConnection,
-  TraceSlice,
-} from '../stats';
+import { Tracer, traceRTCPeerConnection } from '../stats';
+import { StatsTracer } from './StatsTracer';
 
 export type BasePeerConnectionOpts = {
   sfuClient: StreamSfuClient;
@@ -49,9 +43,10 @@ export abstract class BasePeerConnection {
   protected isIceRestarting = false;
   private isDisposed = false;
 
-  protected readonly tracer?: Tracer;
-  private traceStatsInterval?: ReturnType<typeof setInterval>;
-  private costOverrides?: Map<TrackType, number>;
+  protected trackIdToTrackType = new Map<string, TrackType>();
+
+  readonly tracer?: Tracer;
+  readonly stats: StatsTracer;
   private readonly subscriptions: (() => void)[] = [];
   private unsubscribeIceTrickle?: () => void;
 
@@ -81,13 +76,13 @@ export abstract class BasePeerConnection {
       logTag,
     ]);
     this.pc = new RTCPeerConnection(connectionConfig);
+    this.stats = new StatsTracer(this.pc, peerType, this.trackIdToTrackType);
     if (enableTracing) {
       const tag = `${logTag}-${peerType === PeerType.SUBSCRIBER ? 'sub' : 'pub'}`;
       this.tracer = new Tracer(tag);
       this.tracer.trace('clientDetails', clientDetails);
       this.tracer.trace('create', connectionConfig);
       traceRTCPeerConnection(this.pc, this.tracer.trace);
-      this.startStatsTracing();
     }
     this.pc.addEventListener('icecandidate', this.onIceCandidate);
     this.pc.addEventListener('icecandidateerror', this.onIceCandidateError);
@@ -97,6 +92,10 @@ export abstract class BasePeerConnection {
     );
     this.pc.addEventListener('icegatheringstatechange', this.onIceGatherChange);
     this.pc.addEventListener('signalingstatechange', this.onSignalingChange);
+    this.pc.addEventListener(
+      'connectionstatechange',
+      this.onConnectionStateChange,
+    );
   }
 
   /**
@@ -107,7 +106,6 @@ export abstract class BasePeerConnection {
     this.isDisposed = true;
     this.detachEventHandlers();
     this.pc.close();
-    clearInterval(this.traceStatsInterval);
     this.tracer?.dispose();
   }
 
@@ -134,81 +132,6 @@ export abstract class BasePeerConnection {
    * Performs an ICE restart on the `RTCPeerConnection`.
    */
   protected abstract restartIce(): Promise<void>;
-
-  /**
-   * Creates performance stats from the `RTCPeerConnection`.
-   *
-   * @param previousStats the previously collected stats.
-   * @param currentStats the current stats.
-   * @param lastPerformanceStats the last performance stats.
-   * @param iteration the iteration, used for calculating averages.
-   */
-  protected abstract getPerformanceStats(
-    previousStats: Record<string, RTCStats>,
-    currentStats: Record<string, RTCStats>,
-    lastPerformanceStats: PerformanceStats[],
-    iteration: number,
-  ): PerformanceStats[];
-
-  /**
-   * Collects tracing stats from the `RTCPeerConnection`.
-   */
-  private startStatsTracing = () => {
-    let iteration = 1;
-    let prev: Record<string, RTCStats> = {};
-
-    const collectTraceStats = async () => {
-      if (!this.tracer) return;
-      try {
-        const stats = await this.pc.getStats();
-        const now = toObject(stats);
-        const pastStats = this.tracer.getPerformanceStats(this.peerType);
-        const performanceStats = this.getPerformanceStats(
-          prev,
-          now,
-          pastStats || [],
-          iteration++,
-        );
-        if (this.costOverrides) {
-          for (const s of performanceStats) {
-            const override = this.costOverrides.get(s.trackType);
-            if (override !== undefined) {
-              // override the average encode/decode time with the provided cost.
-              // format: [override].[original-encode-time]
-              s.avgFrameTimeMs = override + (s.avgFrameTimeMs || 0) / 1000;
-            }
-          }
-        }
-        this.tracer.setPerformanceStats(this.peerType, performanceStats);
-        this.tracer.trace('getstats', deltaCompression(prev, now));
-        prev = now;
-      } catch (err) {
-        this.tracer.trace('getstatsOnFailure', (err as Error).toString());
-      }
-    };
-
-    this.traceStatsInterval = setInterval(() => {
-      void collectTraceStats();
-    }, 8000);
-
-    this.pc.addEventListener('connectionstatechange', () => {
-      const state = this.pc.connectionState;
-      this.logger('debug', `Connection state changed`, state);
-      if (state === 'connected' || state === 'failed') {
-        void collectTraceStats();
-      }
-    });
-  };
-
-  /**
-   * Sets the performance cost for the given track type.
-   *
-   * @internal don't use this method outside the SDK.
-   */
-  setCost = (cost: number, trackType = TrackType.VIDEO) => {
-    if (!this.costOverrides) this.costOverrides = new Map();
-    this.costOverrides.set(trackType, cost);
-  };
 
   /**
    * Handles events synchronously.
@@ -268,10 +191,10 @@ export abstract class BasePeerConnection {
   };
 
   /**
-   * Returns the current tracing buffer.
+   * Maps the given track ID to the corresponding track type.
    */
-  getTrace = (): TraceSlice | undefined => {
-    return this.tracer?.take();
+  getTrackType = (trackId: string): TrackType | undefined => {
+    return this.trackIdToTrackType.get(trackId);
   };
 
   /**
@@ -306,6 +229,23 @@ export abstract class BasePeerConnection {
       return JSON.stringify({ ...candidate, usernameFragment });
     }
     return JSON.stringify(candidate.toJSON());
+  };
+
+  /**
+   * Handles the ConnectionStateChange event.
+   */
+  private onConnectionStateChange = async () => {
+    const state = this.pc.connectionState;
+    this.logger('debug', `Connection state changed`, state);
+    if (!this.tracer) return;
+    if (state === 'connected' || state === 'failed') {
+      try {
+        const stats = await this.stats.get();
+        this.tracer.trace('getstats', stats.delta);
+      } catch (err) {
+        this.tracer.trace('getstatsOnFailure', (err as Error).toString());
+      }
+    }
   };
 
   /**
