+++ packages/client/src/stats/rtc/Tracer.ts
@@ -1,10 +1,7 @@
 import type { Trace, TraceRecord } from './types';
-import { PeerType, PerformanceStats } from '../../gen/video/sfu/models/models';
 
 export type TraceSlice = {
   snapshot: TraceRecord[];
-  encodeStats: PerformanceStats[] | undefined;
-  decodeStats: PerformanceStats[] | undefined;
   rollback: () => void;
 };
 
@@ -12,7 +9,6 @@ export class Tracer {
   private buffer: TraceRecord[] = [];
   private enabled = true;
   private readonly id: string | null;
-  private readonly performanceStats = new Map<PeerType, PerformanceStats[]>();
 
   constructor(id: string | null) {
     this.id = id;
@@ -29,21 +25,11 @@ export class Tracer {
     this.buffer.push([tag, this.id, data, Date.now()]);
   };
 
-  getPerformanceStats = (peerType: PeerType) => {
-    return this.performanceStats.get(peerType);
-  };
-
-  setPerformanceStats = (peerType: PeerType, stats: PerformanceStats[]) => {
-    this.performanceStats.set(peerType, stats);
-  };
-
   take = (): TraceSlice => {
     const snapshot = this.buffer;
     this.buffer = [];
     return {
       snapshot,
-      encodeStats: this.getPerformanceStats(PeerType.PUBLISHER_UNSPECIFIED),
-      decodeStats: this.getPerformanceStats(PeerType.SUBSCRIBER),
       rollback: () => {
         this.buffer.unshift(...snapshot);
       },
@@ -52,6 +38,5 @@ export class Tracer {
 
   dispose = () => {
     this.buffer = [];
-    this.performanceStats.clear();
   };
 }
